+++ aws/lambda_functions/etl_task_sql/test/runsam.sh
@@ -1,1 +1,1 @@
-sam local invoke "SQL" -e sam_local_event.json
+sam local invoke "SQL" -e sam_event.json
