+++ aws/lambda_functions/etl_task_table_copy/test/runsam.sh
