+++ aws/lambda_functions/etl_task_table_copy/test/sam_event.json
@@ -2,57 +2,24 @@
   "JobIsGo": true,
   "ETLJob": {
     "name": "testtable2",
-    "depends": [
-      "testtable"
-    ],
+    "depends": [],
     "path": "store/assets/testtable/testtable.json",
     "run_group": "tues_night",
     "etl_tasks": [
       {
         "type": "table_copy",
         "source_location": {
-          "connection": "bedrock-googlesheets",
-          "spreadsheetid": "1TsJqzi1n3G8w4eeYIyx_boeHbjiJ1ZGGOuAFuW_j81Q",
-          "range": "Sheet1!A2:B6"
+          "connection": "munis/munprod/mssqlgisadmin",
+          "schemaname": "amd",
+          "tablename": "ad_info"
         },
         "target_location": {
-          "connection": "bedrock-googlesheets",
-          "spreadsheetid": "1TsJqzi1n3G8w4eeYIyx_boeHbjiJ1ZGGOuAFuW_j81Q",
-          "range": "Sheet2!A2:B",
-          "append": true
-        },
-        "active": true
-      },
-      {
-        "type": "table_copy",
-        "source_location": {
-          "connection": "localpg2",
-          "schemaname": "public",
-          "tablename": "testtable"
-        },
-        "target_location": {
-          "connection": "localpg1",
-          "schemaname": "public",
-          "tablename": "testtable"
-        },
-        "active": true
-      },
-      {
-        "type": "table_copy",
-        "source_location": {
-          "connection": "localpg1",
-          "schemaname": "public",
-          "tablename": "tst"
-        },
-        "target_location": {
-          "connection": "bedrock-googlesheets",
-          "spreadsheetid": "1TsJqzi1n3G8w4eeYIyx_boeHbjiJ1ZGGOuAFuW_j81Q",
-          "range": "Sheet2!A2:B",
-          "append": false
+          "connection": "pubrecdb1/mdastore1/dbadmin",
+          "schemaname": "internal",
+          "tablename": "ad_info"
         },
         "active": true
       }
-
     ]
   },
   "TaskIndex": 0,
