--- tests/test_secrets_manager/.gitignore
@@ -1,8 +1,0 @@
-package/
-function.zip
-**/.terraform*
-*.tfstate.*
-*.tfstate
-
-
-
