--- tests/test_secrets_manager/ca.tfvars
@@ -1,4 +1,0 @@
-aws_profile         = "custom-terraform"
-region              = "us-east-1"
-subnet_ids          = ["subnet-00e55df750014753d", "subnet-0c119b605ff498f3b"]
-security_group_ids  = ["sg-076e12ba2a9012944"]
