--- tests/test_secrets_manager/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "test_connections",
-  "version": "1.0.0",
-  "description": "",
-  "main": "handler.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "^2.1119.0"
-  }
-}
