--- tests/test_secrets_manager/variables.tf
@@ -1,19 +1,0 @@
-variable "region" {
-  type          = string
-  description   = "Region in which to create resources"
-}
-
-variable "aws_profile" {
-  type          = string
-  description   = "AWS User Profile to use"
-}
-
-variable "subnet_ids" {
-  type          = list(string)
-  description   = "Array of subnet ids"
-}
-
-variable "security_group_ids" {
-  type          = list(string)
-  description   = "Array of security_group_ids"
-}
